Copy rules files when creating project
